# docs/api/openapi.json
{
  "components" : {
    "responses" : {
      "BadRequest" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "The request is malformed or fails validation; `fields` names the offending fields, never their values."
      },
      "Conflict" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "A conflict: duplicate, last administrator, or a concurrent change that survived the retries."
      },
      "Forbidden" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "The caller's role does not allow this (decided by the use case)."
      },
      "NotFound" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "Not found - also what an id of another association gets."
      },
      "RuleViolation" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "A business rule refuses it; `message` says which."
      },
      "TooManyRequests" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "Too many requests; see `Retry-After`."
      },
      "Unauthenticated" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "No valid access token (missing, expired, forged, or the account is no longer active)."
      }
    },
    "schemas" : {
...
      "ApiError" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "fields" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          }
        }
      },
...
      "CreateTrainingGroupRequest" : {
        "properties" : {
          "acceptedLevelIds" : {
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "maxItems" : 20,
            "minItems" : 1,
            "type" : "array",
            "uniqueItems" : true
          },
          "capacity" : {
            "format" : "int32",
            "maximum" : 1000,
            "minimum" : 1,
            "type" : "integer"
          },
          "coachId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "schedule" : {
            "items" : {
              "$ref" : "#/components/schemas/WeeklySlotRequest"
            },
            "maxItems" : 14,
            "minItems" : 1,
            "type" : "array"
          },
          "venueId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "acceptedLevelIds", "capacity", "coachId", "name", "schedule", "venueId" ],
        "type" : "object"
      },
      "EditTrainingGroupRequest" : {
        "properties" : {
          "acceptedLevelIds" : {
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "maxItems" : 20,
            "minItems" : 1,
            "type" : "array",
            "uniqueItems" : true
          },
          "capacity" : {
            "format" : "int32",
            "maximum" : 1000,
            "minimum" : 1,
            "type" : "integer"
          },
          "coachId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "schedule" : {
            "items" : {
              "$ref" : "#/components/schemas/WeeklySlotRequest"
            },
            "maxItems" : 14,
            "minItems" : 1,
            "type" : "array"
          }
        },
        "required" : [ "acceptedLevelIds", "capacity", "coachId", "name", "schedule" ],
        "type" : "object"
      },
...
      "PaymentRecordedResponse" : {
        "properties" : {
          "outstandingCents" : {
            "format" : "int64",
            "type" : "integer"
          },
          "payment" : {
            "$ref" : "#/components/schemas/PaymentResponse"
          },
          "paymentStatus" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "PaymentResponse" : {
        "properties" : {
          "amountCents" : {
            "format" : "int64",
            "type" : "integer"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "method" : {
            "type" : "string"
          },
          "paidOn" : {
            "format" : "date",
            "type" : "string"
          },
          "recordedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "recordedBy" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reversalOf" : {
            "format" : "uuid",
            "type" : "string"
          },
          "subscriptionId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RecordPaymentRequest" : {
        "properties" : {
          "amountCents" : {
            "exclusiveMinimum" : 0,
            "format" : "int64",
            "maximum" : 100000000,
            "type" : "integer"
          },
          "method" : {
            "enum" : [ "CASH", "TRANSFER", "MB_WAY" ],
            "type" : "string"
          },
          "paidOn" : {
            "format" : "date",
            "type" : "string"
          }
        },
        "required" : [ "amountCents", "method", "paidOn" ],
        "type" : "object"
      },
...
      "SlotView" : {
        "properties" : {
          "dayOfWeek" : {
            "type" : "string"
          },
          "durationMinutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "startTime" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SubscriptionPaymentEntryResponse" : {
        "properties" : {
          "endDate" : {
            "format" : "date",
            "type" : "string"
          },
          "memberId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "memberName" : {
            "type" : "string"
          },
          "paymentStatus" : {
            "type" : "string"
          },
          "planId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "priceCents" : {
            "format" : "int64",
            "type" : "integer"
          },
          "startDate" : {
            "format" : "date",
            "type" : "string"
          },
          "subscriptionId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SubscriptionResponse" : {
        "properties" : {
          "creditsUsed" : {
            "format" : "int32",
            "type" : "integer"
          },
          "endDate" : {
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "memberId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "paymentStatus" : {
            "type" : "string"
          },
          "planId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "priceCents" : {
            "format" : "int64",
            "type" : "integer"
          },
          "startDate" : {
            "format" : "date",
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TrainingGroupResponse" : {
        "properties" : {
          "acceptedLevelIds" : {
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "type" : "array"
          },
          "capacity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "coachId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "schedule" : {
            "items" : {
              "$ref" : "#/components/schemas/SlotView"
            },
            "type" : "array"
          },
          "status" : {
            "type" : "string"
          },
          "venueId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "WeeklySlotRequest" : {
        "properties" : {
          "dayOfWeek" : {
            "enum" : [ "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY" ],
            "type" : "string"
          },
          "durationMinutes" : {
            "format" : "int32",
            "maximum" : 240,
            "minimum" : 1,
            "type" : "integer"
          },
          "startTime" : {
            "format" : "time-local",
            "type" : "string"
          }
        },
        "required" : [ "dayOfWeek", "durationMinutes", "startTime" ],
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "The short-lived access token of POST /auth/login or /auth/refresh.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Session-by-session booking for amateur indoor volleyball associations. All paths are under /api/v1. The tenant\n(association) and the caller always come from the bearer token, never from a path, query or body; an id that\nbelongs to another association answers exactly like an id that does not exist (404). Every failure has the body\ndescribed by ApiError. Times are ISO-8601 instants (UTC); weekly schedules and dates are Europe/Lisbon local time.\nMoney is in cents.",
    "title" : "RegiVolley API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/subscriptions" : {
      "get" : {
        "operationId" : "subscriptionList",
        "parameters" : [ {
          "in" : "query",
          "name" : "paymentStatus",
          "required" : true,
          "schema" : {
            "enum" : [ "PENDING", "PAID", "OVERDUE" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "endingFrom",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "endingTo",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/SubscriptionPaymentEntryResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "subscription-controller" ]
      }
    },
    "/api/v1/subscriptions/export" : {
      "get" : {
        "operationId" : "subscriptionExport",
        "parameters" : [ {
          "in" : "query",
          "name" : "paymentStatus",
          "required" : true,
          "schema" : {
            "enum" : [ "PENDING", "PAID", "OVERDUE" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "endingFrom",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "endingTo",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "text/csv" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "subscription-controller" ]
      }
    },
    "/api/v1/subscriptions/{subscriptionId}/overdue-marking" : {
      "post" : {
        "operationId" : "subscriptionOverdue",
        "parameters" : [ {
          "in" : "path",
          "name" : "subscriptionId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SubscriptionResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "subscription-controller" ]
      }
    },
    "/api/v1/subscriptions/{subscriptionId}/payments" : {
      "post" : {
        "operationId" : "subscriptionPay",
        "parameters" : [ {
          "in" : "path",
          "name" : "subscriptionId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecordPaymentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentRecordedResponse"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "subscription-controller" ]
      }
    },
    "/api/v1/training-groups" : {
      "post" : {
        "operationId" : "trainingGroupCreate",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTrainingGroupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrainingGroupResponse"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "training-group-controller" ]
      }
    },
    "/api/v1/training-groups/{groupId}" : {
      "put" : {
        "operationId" : "trainingGroupEdit",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EditTrainingGroupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrainingGroupResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "training-group-controller" ]
      }
    },
    "/api/v1/training-groups/{groupId}/archival" : {
      "post" : {
        "operationId" : "trainingGroupArchive",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrainingGroupResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "training-group-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}